requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==24.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.0.0
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cached-property==1.5.2
cachetools==5.5.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.8
cloudpickle==2.2.1
comm==0.1.4
config==0.5.1
ConfigArgParse==1.7.1
cupy==11.6.0
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.5
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.3.0
fastjsonschema==2.21.2
fastrlock==0.8.3
filelock==3.12.2
Flask==2.2.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.1.0
fvcore==0.1.5.post20221221
google-auth==2.40.3
google-auth-oauthlib==0.4.6
grpcio==1.62.3
huggingface-hub==0.16.4
idna @ file:///croot/idna_1666125576474/work
ignite==1.1.0
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
json5==0.9.16
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
natsort==8.4.0
nbclassic==1.2.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==2.6.3
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.17.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==24.0
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.5
pathlib==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygit2==1.10.1
Pygments==2.17.2
pyparsing==3.1.4
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.9.0.post0
python-json-logger==3.0.1
pytorch-ignite==0.5.2
pytz==2025.2
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==26.2.1
qudida==0.0.4
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==2.0.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.4.1
submitit==1.4.6
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchcontrib==0.0.2
torchmetrics==0.11.4
torchvision==0.14.1+cu117
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1718728347128/work
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.14
wrapt==1.16.0
y-py==0.6.2
yacs==0.1.8
ypy-websocket==0.8.4
zipp==3.15.0
zope.interface==6.4.post2
